waltersJeffreyIteration1
In this Milestone, I added a Boss Health Bar, added Sound to the player's shooter, and added a special effect to the boss and enemy dying. I realized that the boss, how has more health, could use a health bar in order to let the players know how many shots it will take to kill her. I can’t think of one game that doesn’t have sound, so I placed an audio source and a free shooting sound effect to the players weapon. My favorite out of the milestone is the special effects that I included on the enemies for it gave it a good touch. I am not done yet, but I am getting closer to the completion of my game. I’m so Excited!