waltersJeffreyFinal
In the game, please start with the cutscene, it will give you the beginning of the story. My decision (resolution) for this project is to take your time on the game. It is ok to lose the first try. I don’t think that there is anything that you need to check besides the things that are said to be fixed in the video. I personally hope the audio is not too loud.